pip-requirements-12.txt
numpy>=1.24
librosa>=0.10
tqdm
soundfile
pandas
pyyaml
matplotlib
num2words
resampy
tensorflow
scikit-learn
joblib
